chore: Refractor org navbar team setup to another file
